requirements.txt
torch
dgl
numpy
pandas
scipy
biopython
transformers
fair-esm
matplotlib
tqdm
tmtools
